# api/swagger.yaml
basePath: /web
definitions:
...
  web.CloneTransactionRequest:
    properties:
      id:
        example: 42
        type: integer
    type: object
...
  web.EditTransactionRequest:
    properties:
      amount:
        example: 19.9
        type: number
      category:
        example: Grocery
        type: string
      date:
        example: "2026-05-21"
        type: string
      description:
        example: weekly shop
        type: string
      id:
        example: 42
        type: integer
    type: object
  web.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  web.SearchTransactionsRequest:
    properties:
      amountMax:
        example: 100
        type: number
      amountMin:
        example: 5
        type: number
      category:
        example: Grocery
        type: string
      dateFrom:
        example: "2026-01-01"
        type: string
      dateTo:
        example: "2026-05-31"
        type: string
      limit:
        example: 50
        type: integer
      offset:
        example: 0
        type: integer
      query:
        example: coffee
        type: string
      type:
        example: Expense
        type: string
    type: object
  web.SearchTransactionsResponse:
    properties:
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
      transactions:
        items:
          $ref: '#/definitions/web.TransactionDTO'
        type: array
    type: object
...
  web.TransactionDTO:
    properties:
      amount:
        type: number
      category:
        type: string
      date:
        type: string
      description:
        type: string
      id:
        type: integer
      type:
        type: string
    type: object
...
info:
  contact: {}
  description: 'Personal finance tracking API. All /api/* endpoints require either
    a session cookie (browser) or an Authorization: Bearer <token> header (machine
    clients).'
  title: Pago API
  version: "1.0"
paths:
...
  /api/transactions/clone:
    post:
      consumes:
      - application/json
      description: Duplicate an existing transaction; the new transaction copies type,
        category, amount, description and currency, with date set to today.
      parameters:
      - description: ID of the transaction to clone
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/web.CloneTransactionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.TransactionDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/web.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Clone transaction
      tags:
      - transactions
...
  /api/transactions/edit:
    patch:
      consumes:
      - application/json
      description: Update one or more fields of an existing transaction. Type cannot
        be changed; category must remain within the same type (Income↔Expense swaps
        are rejected).
      parameters:
      - description: Fields to update; only non-null fields are applied
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/web.EditTransactionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.TransactionDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/web.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Edit transaction (partial)
      tags:
      - transactions
  /api/transactions/export:
    get:
      description: 'Stream a CSV containing all transactions matching the optional
        filter set. Columns: date,type,category,amount,currency,description,created_at,updated_at.'
      parameters:
      - description: Substring match on description (case-insensitive)
        in: query
        name: query
        type: string
      - description: Category filter (\
        in: query
        name: category
        type: string
      - description: 'Transaction type: Income or Expense'
        in: query
        name: type
        type: string
      - description: Inclusive lower bound (YYYY-MM-DD)
        in: query
        name: dateFrom
        type: string
      - description: Inclusive upper bound (YYYY-MM-DD)
        in: query
        name: dateTo
        type: string
      - description: Inclusive lower bound on amount
        in: query
        name: amountMin
        type: number
      - description: Inclusive upper bound on amount
        in: query
        name: amountMax
        type: number
      produces:
      - text/csv
      responses:
        "200":
          description: OK
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Export transactions as CSV
      tags:
      - transactions
...
  /api/transactions/search:
    post:
      consumes:
      - application/json
      description: Search a user's transactions by any combination of text, category,
        type, date range, amount range. Returns paginated results with a total count.
      parameters:
      - description: Filter set
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/web.SearchTransactionsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.SearchTransactionsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Search transactions
      tags:
      - transactions
...
securityDefinitions:
  BearerAuth:
    description: 'Provide the value as: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"